Extract bench to separate file

+++ types.h
@@ -1,3 +1,3 @@
 #include <stdint.h>
 
-typedef int32_t data_t;
+typedef int8_t data_t;
